Initial Commit
Attempt at a lifelogging app based on the “winky” app. All the buttons
on the screen don’t work. Don’t click them.

Problems:
Main issue is that the picture-taking action happens every 30 sec
(currently) and will interrupt anything that is happening when it takes
a picture. So, you can only use your glass for 30-sec at a time before
it goes back to the home screen and/or off.

It’s also not clear if it’s still taking a picture when it goes to
sleep.

Also, all the pictures end up in the main photo pool, which syncs to
google and you end up with a ton of photos.